Rename module: migration_status -> table_migration_status
This module is related to the status of table migration, not the broader UC migration.
