+++ src/databricks/labs/ucx/contexts/application.py
@@ -35,7 +35,7 @@ from databricks.labs.ucx.hive_metastore.grants import (
     ACLMigrator,
 )
 from databricks.labs.ucx.hive_metastore.mapping import TableMapping
-from databricks.labs.ucx.hive_metastore.migration_status import TableMigrationIndex
+from databricks.labs.ucx.hive_metastore.table_migration_status import TableMigrationIndex
 from databricks.labs.ucx.hive_metastore.table_migrate import (
     TableMigrationStatusRefresher,
     TablesMigrator,
