+++ src/databricks/labs/ucx/source_code/linters/context.py
@@ -2,7 +2,7 @@ from typing import cast
 
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.hive_metastore.migration_status import TableMigrationIndex
+from databricks.labs.ucx.hive_metastore.table_migration_status import TableMigrationIndex
 from databricks.labs.ucx.source_code.base import (
     Fixer,
     Linter,
