+++ src/databricks/labs/ucx/source_code/lsp.py
@@ -12,10 +12,10 @@ from urllib.parse import parse_qsl
 from databricks.labs.blueprint.logger import install_logger
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.hive_metastore.migration_status import (
+from databricks.labs.ucx.hive_metastore.table_migration_status import (
     TableMigrationStatus,
 )
-from databricks.labs.ucx.hive_metastore.migration_status import TableMigrationIndex
+from databricks.labs.ucx.hive_metastore.table_migration_status import TableMigrationIndex
 from databricks.labs.ucx.source_code.base import (
     Advice,
     Advisory,
