+++ tests/unit/source_code/conftest.py
@@ -1,9 +1,6 @@
 import pytest
 
-from databricks.labs.ucx.hive_metastore.migration_status import (
-    TableMigrationStatus,
-)
-from databricks.labs.ucx.hive_metastore.migration_status import TableMigrationIndex
+from databricks.labs.ucx.hive_metastore.table_migration_status import TableMigrationIndex, TableMigrationStatus
 from databricks.labs.ucx.source_code.graph import DependencyResolver
 from databricks.labs.ucx.source_code.known import KnownList
 from databricks.labs.ucx.source_code.linters.files import ImportFileResolver, FileLoader
